+++ src/main/java/com/unithon/team15_server/global/university/implement/UniversityDataLoader.java
@@ -3,8 +3,8 @@ package com.unithon.team15_server.global.university.implement;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.unithon.team15_server.global.exception.CustomException;
 import com.unithon.team15_server.global.exception.ErrorCode;
-import com.unithon.team15_server.global.university.AccreditedUniversity;
-import com.unithon.team15_server.global.university.AccreditedUniversityRepository;
+import com.unithon.team15_server.global.university.entity.AccreditedUniversity;
+import com.unithon.team15_server.global.university.repository.AccreditedUniversityRepository;
 import com.unithon.team15_server.global.university.dto.AccreditedUniversityDto;
 import com.unithon.team15_server.global.university.enums.ProgramType;
 import com.unithon.team15_server.global.university.enums.UniversityType;
