+++ src/main/java/com/unithon/team15_server/global/university/repository/AccreditedUniversityRepository.java
@@ -1,5 +1,6 @@
-package com.unithon.team15_server.global.university;
+package com.unithon.team15_server.global.university.repository;
 
+import com.unithon.team15_server.global.university.entity.AccreditedUniversity;
 import com.unithon.team15_server.global.university.enums.ProgramType;
 import org.springframework.data.jpa.repository.JpaRepository;
